Fix routing from push
